pyproject.toml
[tool.poetry]
name = "hw-12"
version = "0.1.0"
description = ""
authors = ["Anton Holovin <54354764+Unfeir@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "hw_12"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = {extras = ["basic"], version = "^0.21.1"}
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.9"
alembic = "^1.10.3"
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
libgravatar = "^1.0.4"
pydantic = {extras = ["email"], version = "^1.10.7"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
